+++ CHANGELOG.md
@@ -17,6 +17,13 @@ This is an example of a brief overview of the _Major_ or _Minor_ version changes
 
 ---
 
+## [2.9] 🍞 Toasts
+
+#### **Changes**
+- Adds `Toast` component to `fds-components`
+- Adds `Toaster` controller component to `fds-components`
+- Adds `CountdownButton` component to `fds-components`
+
 ## [2.8] Material ui theme
 
 #### **Changes**